(change, add) add DTO to show category of a product and changed data types
Product had profit margin set to Float instead of double like the
database had set it up. So every other DTO that used that data type had
to be changed as well.

+++ src/main/java/com/utn/ProgIII/model/Product/Category.java
@@ -19,7 +19,7 @@ public class Category {
 
     @Id
     @GeneratedValue (strategy = GenerationType.IDENTITY)
-    private Integer idCategory;
+    private Long idCategory;
 
     @NotBlank(message = "El nombre de la categoria no puede ser vacio")
     @Length(min = 3, max = 30, message = "El nombre de la categoria debe tener entre 3 y 30 caracteres")
